引入SpotBugs插件，并优化CheckStyle插件
