--- src/main/java/com/github/main.java
@@ -1,106 +1,0 @@
-package com.github;
-
-import org.apache.http.HttpEntity;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.util.EntityUtils;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
-
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-public class main {
-    public static void main(String[] args) throws IOException {
-
-//        待处理链接池
-        List<String> linkPool = new ArrayList<>();
-//        已经处理链接池
-        Set<String> processedLinks = new HashSet<>();
-        linkPool.add("https://sina.cn");
-
-        while (true) {
-            if (linkPool.isEmpty()) {
-                break;
-            }
-
-            // ArrayList从尾部删除更有效率
-            String link = linkPool.remove(linkPool.size() - 1);
-
-            if (link.contains("\\/")) {
-                continue;
-            }
-
-            if (processedLinks.contains(link)) {
-                continue;
-            }
-
-            // 我们只关心News.sina，我们要排除登录页面
-            if (isInterestingLink(link)) {
-                Document doc = httpGetAndParseHtml(link);
-                doc.select("a").stream().map(aTag -> aTag.attr("href")).forEach(linkPool::add);
-                // 假如这是一个新闻的详情页面，就加入数据库，否则，就什么都不做
-                storeIntoDatabaseIfItIsNewsPags(doc);
-                processedLinks.add(link);
-            } else {
-                // 这是我们不感兴趣的，不处理它
-            }
-        }
-    }
-
-    private static void storeIntoDatabaseIfItIsNewsPags(Document doc) {
-        ArrayList<Element> articleTags = doc.select("article");
-        if (!articleTags.isEmpty()) {
-            for (Element articleTag : articleTags) {
-                String title = articleTags.get(0).child(0).text();
-                System.out.println(title);
-            }
-        }
-    }
-
-    private static Document httpGetAndParseHtml(String link) throws IOException {
-        // 这是我们感兴趣的，我们只处理新浪站内的链接
-        CloseableHttpClient httpclient = HttpClients.createDefault();
-
-
-        System.out.println(link);
-        if (link.startsWith("//")) {
-            link = "https" + link;
-            System.out.println(link);
-        }
-
-
-        HttpGet httpGet = new HttpGet(link);
-        httpGet.addHeader("user-agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/80.0.3987.106 Safari/537.36");
-
-        try (CloseableHttpResponse response1 = httpclient.execute(httpGet)) {
-            System.out.println(response1.getStatusLine());
-            HttpEntity entity1 = response1.getEntity();
-            String html = EntityUtils.toString(entity1);
-            return Jsoup.parse(html);
-        }
-    }
-
-    private static boolean isInterestingLink(String link) {
-        return (isNewsPage(link) || isIndexPage(link) && isNotLoginPage(link));
-    }
-
-    private static boolean isNewsPage(String link) {
-        return link.contains("news.sina.cn");
-    }
-
-    private static boolean isIndexPage(String link) {
-        return "https://sina.cn".equals(link);
-    }
-
-    private static boolean isNotLoginPage(String link) {
-        return !link.contains("passport.sina.cn");
-    }
-}
